+++ src/core/rendering/sensor/buffer_opaque.zig
@@ -36,7 +36,7 @@ pub const Opaque = struct {
         }
     }
 
-    pub fn addPixel(self: *Opaque, i: usize, color: Vec4f, weight: f32) void {
+    pub fn addPixel(self: *Opaque, i: u32, color: Vec4f, weight: f32) void {
         const wc = @as(Vec4f, @splat(weight)) * color;
         var value: Vec4f = self.pixels[i].v;
         value += Vec4f{ wc[0], wc[1], wc[2], weight };
@@ -44,7 +44,7 @@ pub const Opaque = struct {
         self.pixels[i].v = value;
     }
 
-    pub fn addPixelAtomic(self: *Opaque, i: usize, color: Vec4f, weight: f32) void {
+    pub fn addPixelAtomic(self: *Opaque, i: u32, color: Vec4f, weight: f32) void {
         const wc = @as(Vec4f, @splat(weight)) * color;
 
         var value = &self.pixels[i];
@@ -54,7 +54,7 @@ pub const Opaque = struct {
         _ = @atomicRmw(f32, &value.v[3], .Add, weight, .Monotonic);
     }
 
-    pub fn splatPixelAtomic(self: *Opaque, i: usize, color: Vec4f, weight: f32) void {
+    pub fn splatPixelAtomic(self: *Opaque, i: u32, color: Vec4f, weight: f32) void {
         const wc = @as(Vec4f, @splat(weight)) * color;
 
         var value = &self.pixels[i];
