+++ src/core/rendering/sensor/buffer_transparent.zig
@@ -43,7 +43,7 @@ pub const Transparent = struct {
         }
     }
 
-    pub fn addPixel(self: *Transparent, i: usize, color: Vec4f, weight: f32) void {
+    pub fn addPixel(self: *Transparent, i: u32, color: Vec4f, weight: f32) void {
         self.pixel_weights[i] += weight;
 
         const wc = @as(Vec4f, @splat(weight)) * color;
@@ -53,7 +53,7 @@ pub const Transparent = struct {
         self.pixels[i].v = value;
     }
 
-    pub fn addPixelAtomic(self: *Transparent, i: usize, color: Vec4f, weight: f32) void {
+    pub fn addPixelAtomic(self: *Transparent, i: u32, color: Vec4f, weight: f32) void {
         _ = @atomicRmw(f32, &self.pixel_weights[i], .Add, weight, .Monotonic);
 
         const wc = @as(Vec4f, @splat(weight)) * color;
@@ -65,7 +65,7 @@ pub const Transparent = struct {
         _ = @atomicRmw(f32, &value.v[3], .Add, wc[3], .Monotonic);
     }
 
-    pub fn splatPixelAtomic(self: *Transparent, i: usize, color: Vec4f, weight: f32) void {
+    pub fn splatPixelAtomic(self: *Transparent, i: u32, color: Vec4f, weight: f32) void {
         const wc = @as(Vec4f, @splat(weight)) * color;
 
         var value = &self.pixels[i];
